refactor -> ScoreBoard class를 다른 패키지로 이동
