+++ src/main/java/Score/ScoreBoard.java
@@ -1,4 +1,4 @@
-package baseball;
+package Score;
 
 import java.util.List;
 import java.util.Objects;
